+++ src/Testcontainers/Configurations/ICustomConfiguration.cs
@@ -1,6 +1,7 @@
 ﻿namespace DotNet.Testcontainers.Configurations
 {
   using System;
+  using System.Text.Json;
   using DotNet.Testcontainers.Images;
   using JetBrains.Annotations;
 
@@ -10,12 +11,28 @@
   internal interface ICustomConfiguration
   {
     /// <summary>
+    /// Gets the Docker config custom configuration.
+    /// </summary>
+    /// <returns>The Docker config custom configuration.</returns>
+    /// <remarks>https://www.testcontainers.org/features/configuration/#customizing-docker-host-detection.</remarks>
+    [CanBeNull]
+    string GetDockerConfig();
+
+    /// <summary>
     /// Gets the Docker host custom configuration.
     /// </summary>
     /// <returns>The Docker host custom configuration.</returns>
     /// <remarks>https://www.testcontainers.org/features/configuration/#customizing-docker-host-detection.</remarks>
     [CanBeNull]
     Uri GetDockerHost();
+
+    /// <summary>
+    /// Gets the Docker registry authentication custom configuration.
+    /// </summary>
+    /// <returns>The Docker authentication custom configuration.</returns>
+    /// <remarks>https://docs.gitlab.com/ee/ci/docker/using_docker_images.html#access-an-image-from-a-private-container-registry.</remarks>
+    [CanBeNull]
+    JsonDocument GetDockerAuthConfig();
 
     /// <summary>
     /// Gets the Ryuk disabled custom configuration.