+++ src/Testcontainers/Configurations/TestcontainersSettings.cs
@@ -20,10 +20,6 @@ namespace DotNet.Testcontainers.Configurations
   {
     private static readonly IDockerImage RyukContainerImage = new DockerImage("testcontainers/ryuk:0.3.4");
 
-    private static readonly ICustomConfiguration PropertiesFileConfiguration = new PropertiesFileConfiguration();
-
-    private static readonly ICustomConfiguration EnvironmentConfiguration = new EnvironmentConfiguration();
-
     private static readonly IDockerEndpointAuthenticationConfiguration DockerEndpointAuthConfig =
       new IDockerEndpointAuthenticationProvider[] { new EnvironmentEndpointAuthenticationProvider(), new NpipeEndpointAuthenticationProvider(), new UnixEndpointAuthenticationProvider() }
         .AsParallel()
@@ -37,14 +33,14 @@ namespace DotNet.Testcontainers.Configurations
     /// </summary>
     [PublicAPI]
     public static bool ResourceReaperEnabled { get; set; }
-      = !PropertiesFileConfiguration.GetRyukDisabled() && !EnvironmentConfiguration.GetRyukDisabled();
+      = !PropertiesFileConfiguration.Instance.GetRyukDisabled() && !EnvironmentConfiguration.Instance.GetRyukDisabled();
 
     /// <summary>
     /// Gets or sets the <see cref="ResourceReaper" /> image.
     /// </summary>
     [PublicAPI]
     public static IDockerImage ResourceReaperImage { get; set; }
-      = PropertiesFileConfiguration.GetRyukContainerImage() ?? EnvironmentConfiguration.GetRyukContainerImage() ?? RyukContainerImage;
+      = PropertiesFileConfiguration.Instance.GetRyukContainerImage() ?? EnvironmentConfiguration.Instance.GetRyukContainerImage() ?? RyukContainerImage;
 
     /// <summary>
     /// Gets or sets the <see cref="ResourceReaper" /> public host port.
@@ -68,7 +64,7 @@ namespace DotNet.Testcontainers.Configurations
     [PublicAPI]
     [CanBeNull]
     public static string HubImageNamePrefix { get; set; }
-      = PropertiesFileConfiguration.GetHubImageNamePrefix() ?? EnvironmentConfiguration.GetHubImageNamePrefix();
+      = PropertiesFileConfiguration.Instance.GetHubImageNamePrefix() ?? EnvironmentConfiguration.Instance.GetHubImageNamePrefix();
 
     /// <summary>
     /// Gets or sets the logger.