+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 520 Add MariaDB module (@renemadsen)
 - 528 Do not require the Docker host configuration (`DockerEndpointAuthConfig`) on `TestcontainersSettings` initialization
 - 538 Support optional username and password in MongoDB connection string (@the-avid-engineer)
+- 540 Add Docker registry authentication provider for `DOCKER_AUTH_CONFIG` environment variable (@vova-lantsov-dev)
 - 541 Allow MsSqlTestcontainerConfiguration custom database names (@enginexon)
 
 ### Fixed